Refactor customer-resource-deselection to use BigTest page objects and assertions
